+++ src/main/java/com/quasiris/qsf/pipeline/filter/PagingFilter.java
@@ -1,5 +1,6 @@
 package com.quasiris.qsf.pipeline.filter;
 
+import com.quasiris.qsf.paging.PagingBuilder;
 import com.quasiris.qsf.pipeline.PipelineContainer;
 import com.quasiris.qsf.response.Page;
 import com.quasiris.qsf.response.Paging;
@@ -35,35 +36,10 @@ public class PagingFilter extends AbstractFilter {
             rows = 10;
         }
 
-        Paging paging = buildPaging(searchResult.getTotal(), currentPage, rows);
+        Paging paging = PagingBuilder.buildPaging(searchResult.getTotal(), currentPage, rows);
         searchResult.setPaging(paging);
 
         return pipelineContainer;
 
-    }
-
-    private static Paging buildPaging(Long total, Integer currentPage, Integer rows) {
-        Long pageCount = ((total / rows) +1);
-
-        Paging paging = new Paging();
-        paging.setCurrentPage(currentPage);
-        paging.setPageCount(pageCount.intValue());
-        paging.setFirstPage(createPage(1, currentPage));
-        paging.setLastPage(createPage(pageCount.intValue(), currentPage));
-
-
-        paging.setNextPage(createPage(Math.min(pageCount.intValue(), (currentPage+1)), currentPage));
-        paging.setPreviousPage(createPage(Math.max(1, (currentPage-1)), currentPage));
-
-        paging.setRows(rows);
-
-        return paging;
-    }
-
-    private static Page createPage(int number, Integer currentPage) {
-        Page page = new Page();
-        page.setCurrentPage(number == currentPage);
-        page.setNumber(number);
-        return page;
     }
 }