+++ src/main/java/com/quasiris/qsf/query/Slider.java
@@ -1,0 +1,116 @@
+package com.quasiris.qsf.query;
+
+public class Slider {
+
+    private String id;
+    private String name;
+
+    private String type = "stats";
+
+    /**
+     * Getter for property 'id'.
+     *
+     * @return Value for property 'id'.
+     */
+    public String getId() {
+        return id;
+    }
+
+    /**
+     * Setter for property 'id'.
+     *
+     * @param id Value to set for property 'id'.
+     */
+    public void setId(String id) {
+        this.id = id;
+    }
+
+    /**
+     * Getter for property 'name'.
+     *
+     * @return Value for property 'name'.
+     */
+    public String getName() {
+        return name;
+    }
+
+    /**
+     * Setter for property 'name'.
+     *
+     * @param name Value to set for property 'name'.
+     */
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    /**
+     * Getter for property 'type'.
+     *
+     * @return Value for property 'type'.
+     */
+    public String getType() {
+        return type;
+    }
+
+    /**
+     * Setter for property 'type'.
+     *
+     * @param type Value to set for property 'type'.
+     */
+    public void setType(String type) {
+        this.type = type;
+    }
+
+
+    @Override
+    public String toString() {
+        return "Slider{" +
+                "id='" + id + '\'' +
+                ", name='" + name + '\'' +
+                ", type='" + type + '\'' +
+                '}';
+    }
+
+    public static final class Builder {
+        private String id;
+        private String name;
+        private String type = "stats";
+        private Integer size;
+
+        private Builder() {
+        }
+
+        public static Builder create() {
+            return new Builder();
+        }
+
+        public Builder id(String id) {
+            this.id = id;
+            return this;
+        }
+
+        public Builder name(String name) {
+            this.name = name;
+            return this;
+        }
+
+        public Builder type(String type) {
+            this.type = type;
+            return this;
+        }
+
+        public Builder size(Integer size) {
+            this.size = size;
+            return this;
+        }
+
+        public Slider build() {
+            Slider slider = new Slider();
+            slider.setId(id);
+            slider.setName(name);
+            slider.setType(type);
+            return slider;
+        }
+    }
+
+}